requirements.txt
django
pillow
qrcode
djangorestframework
openai
jsonfield
dnspython
# django-cors-headers